BUGFIX/Improvement: Now using relative paths for API calls.
In viewer, architect, campus_mode, we use relative paths to resolve the
API calls.

This means the configuration does not have to change prior to the
compilation depending whether we want to build for:
localhost, develop, or production.

It also covers whether http/https is used, and also the port number
(e.g. localhost uses port 9000)